Fix the keyring that we use for testing
The original keyring that we generated was not created properly -- it
had trust issues (ha) and did not have a proper copy of its own private
key.

This keyring was generated via the command-line gpg tools, which do a
much more thorough job. The only downside is that they only seem to work
on the .gnupg directory that's in your home directory. So you need to
move your own keyring aside to properly create a new keyring like this
one.